+++ tables/tests/test_attributes.py
@@ -3,7 +3,7 @@
 import datetime
 import sys
 import warnings
-from distutils.version import LooseVersion
+from packaging.version import Version
 
 import numpy as np
 
@@ -1691,7 +1691,7 @@ class NoSysAttrsClose(NoSysAttrsTestCase):
 class CompatibilityTestCase(common.TestFileMixin, common.PyTablesTestCase):
     h5fname = common.test_filename('issue_368.h5')
 
-    @common.unittest.skipIf(LooseVersion(np.__version__) < '1.9.0',
+    @common.unittest.skipIf(Version(np.__version__) < Version('1.9.0'),
                             'requires numpy >= 1.9')
     def test_pickled_unicode_attrs(self):
         # See also gh-368 and https://github.com/numpy/numpy/issues/4879.