+++ tables/tests/test_create.py
@@ -14,6 +14,7 @@ import hashlib
 import tempfile
 import warnings
 from pathlib import Path
+from packaging.version import Version
 
 import numpy as np
 
@@ -1758,7 +1759,7 @@ class DefaultDriverTestCase(common.TempFileMixin, common.PyTablesTestCase):
         self.assertEqual(root.table2.cols.var2.dtype, tb.FloatCol().dtype)
 
 
-@common.unittest.skipIf(common.hdf5_version < "1.8.9",
+@common.unittest.skipIf(common.hdf5_version < Version("1.8.9"),
                         "requires HDF5 >= 1.8,9")
 class Sec2DriverTestCase(DefaultDriverTestCase):
     DRIVER = "H5FD_SEC2"
@@ -1770,7 +1771,7 @@ class Sec2DriverTestCase(DefaultDriverTestCase):
         self.assertEqual([i for i in image[:4]], [137, 72, 68, 70])
 
 
-@common.unittest.skipIf(common.hdf5_version < "1.8.9",
+@common.unittest.skipIf(common.hdf5_version < Version("1.8.9"),
                         "requires HDF5 >= 1.8,9")
 class StdioDriverTestCase(DefaultDriverTestCase):
     DRIVER = "H5FD_STDIO"
@@ -1782,7 +1783,7 @@ class StdioDriverTestCase(DefaultDriverTestCase):
         self.assertEqual([i for i in image[:4]], [137, 72, 68, 70])
 
 
-@common.unittest.skipIf(common.hdf5_version < "1.8.9",
+@common.unittest.skipIf(common.hdf5_version < Version("1.8.9"),
                         "requires HDF5 >= 1.8,9")
 class CoreDriverTestCase(DefaultDriverTestCase):
     DRIVER = "H5FD_CORE"
@@ -2011,7 +2012,7 @@ class CoreDriverNoBackingStoreTestCase(common.PyTablesTestCase):
         # ensure that there is no change on the file on disk
         self.assertEqual(hexdigest, self._get_digest(self.h5fname))
 
-    @common.unittest.skipIf(common.hdf5_version < "1.8.9",
+    @common.unittest.skipIf(common.hdf5_version < Version("1.8.9"),
                             'HDF5 >= "1.8.9" required')
     def test_get_file_image(self):
         self.h5file = tb.open_file(self.h5fname, mode="w",
@@ -2154,7 +2155,7 @@ class StreamDriverTestCase(NotSpportedDriverTestCase):
     DRIVER = "H5FD_STREAM"
 
 
-@common.unittest.skipIf(common.hdf5_version < "1.8.9",
+@common.unittest.skipIf(common.hdf5_version < Version("1.8.9"),
                         'HDF5 >= "1.8.9" required')
 class InMemoryCoreDriverTestCase(common.PyTablesTestCase):
     DRIVER = "H5FD_CORE"